youtube_post_agent/requirements.txt
moviepy>=1.0.3
Pillow>=10.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
whisper>=1.1.1
numpy>=1.24.0
requests>=2.31.0